Add initial project structure and configuration
This commit introduces the initial skeleton of the project, including setup and configuration files, environment variables example file, and python packages in setup.py. It also includes the foundation for event handling with the creation of 'music' and 'music_tests' packages, and the addition of assets and resources related to artists and concerts. The readme file was also added to guide project setup and execution.